Add explicit version
Instead of context being passed into start methood of a component
and component then taking dependencies out of it, start method
is called by passing components as arguments followed by callback.

If there are fewer arguments than declared dependencies, the last
argument is assumed to be callback and we pass as many dependencies
as possible in remaining arguments.